Missing source code changes not published in the initial release.
The server ran a copy with fixes that made us believe the clients were
working and the server didn't need fixes, as such it was missed that
Trim commands were actually added in the server copy of the source code.
The Trim code is vital to make some multichain commands work,
furthermore StatementController has had a fix, it turned out the 8 day
delay between recommendations for new users (or users with below initial
trust) was made with the wrong operator "<" instead of ">" when
comparing dates.